ENH: Add the vtkSegmentationModifier helper class
Previously handling of shared layers took place in qSlicerSegmentEditorAbstractEffect::modifySegmentByLabelmap.
This process now takes place in vtkSegmentationModifier::ModifyBinaryLabelmap(), which provides access to labelmap modification and layer management at the PolySeg level.
This commit also fixes a number of compiler warnings in Segmentations.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=28760
git-svn-id: http://svn.slicer.org/Slicer4/trunk@28760 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLSegmentationStorageNode.cxx
@@ -807,7 +807,7 @@ int vtkMRMLSegmentationStorageNode::ReadBinaryLabelmapRepresentation(vtkMRMLSegm
     }
 
   // Add the created segments to the segmentation
-  for (int segmentIndex = 0; segmentIndex < segments.size(); ++segmentIndex)
+  for (int segmentIndex = 0; segmentIndex < static_cast<int>(segments.size()); ++segmentIndex)
     {
     vtkSegment* currentSegment = segments[segmentIndex];
     if (!currentSegment)
